fix on FreeBSD

+++ termios/termios_bsd.go
@@ -73,10 +73,10 @@ func Tcflush(fd, which uintptr) error {
 }
 
 // Cfgetispeed returns the input baud rate stored in the termios structure.
-func Cfgetispeed(attr *syscall.Termios) uint64 { return attr.Ispeed }
+func Cfgetispeed(attr *syscall.Termios) uint32 { return attr.Ispeed }
 
 // Cfgetospeed returns the output baud rate stored in the termios structure.
-func Cfgetospeed(attr *syscall.Termios) uint64 { return attr.Ospeed }
+func Cfgetospeed(attr *syscall.Termios) uint32 { return attr.Ospeed }
 
 // Tiocinq returns the number of bytes in the input buffer.
 func Tiocinq(fd uintptr, argp *int) error {
